Separate pusher parsers into their own file
