+++ src/internal/pusher/pusher.ts
@@ -3,8 +3,8 @@ import { handleRewardRedeemedEvent } from "../../events/reward-redeemed-event";
 import { handleStreamHostedEvent } from "../../events/stream-hosted-event";
 import { integration } from "../../integration";
 import { logger } from "../../main";
-import { ChatMessage, KickUser, RewardRedeemedEvent, StreamHostedEvent } from "../../shared/types";
-import { parseDate } from "../util";
+import { KickUser } from "../../shared/types";
+import { parseChatMessageEvent, parseRewardRedeemedEvent, parseStreamHostedEvent } from "./pusher-parsers";
 
 const Pusher = require('pusher-js');
 
@@ -93,13 +93,13 @@ export class KickPusher {
         try {
             switch (event) {
                 case 'App\\Events\\ChatMessageEvent':
-                    await handleChatMessageSentEvent(this.parseChatMessageEvent(data), 2); // Delay by 2 seconds in hopes webhook arrives first
+                    await handleChatMessageSentEvent(parseChatMessageEvent(data, this.getBroadcaster()), 2); // Delay by 2 seconds in hopes webhook arrives first
                     break;
                 case 'App\\Events\\StreamHostedEvent':
-                    await handleStreamHostedEvent(this.parseStreamHostedEvent(data));
+                    await handleStreamHostedEvent(parseStreamHostedEvent(data));
                     break;
                 case 'RewardRedeemedEvent':
-                    await handleRewardRedeemedEvent(this.parseRewardRedeemedEvent(data));
+                    await handleRewardRedeemedEvent(parseRewardRedeemedEvent(data));
                     break;
                 case 'pusher:subscription_succeeded':
                     logger.info("Pusher subscribed successfully to chatroom events.");
@@ -111,73 +111,6 @@ export class KickPusher {
         } catch (error) {
             logger.error(`Error handling Pusher chatroom event: ${event}, error: ${error}`);
         }
-    }
-
-    private parseChatMessageEvent(data: any): ChatMessage {
-        const d = data as ChatMessageEvent;
-        return {
-            messageId: d.id,
-            broadcaster: this.getBroadcaster(),
-            sender: {
-                userId: d.sender.id.toString(),
-                username: d.sender.username,
-                displayName: d.sender.username,
-                profilePicture: '', // Not provided in event
-                isVerified: false, // Worth checking?
-                channelSlug: d.sender.slug,
-                identity: {
-                    usernameColor: d.sender.identity.color,
-                    badges: d.sender.identity.badges.map(b => ({
-                        text: b.text,
-                        type: b.type,
-                        count: b.count
-                    }))
-                }
-            },
-            content: d.content,
-            createdAt: parseDate(d.created_at),
-            repliesTo: d.metadata && d.metadata.original_message && d.metadata.original_sender ? {
-                messageId: d.metadata.original_message.id,
-                content: d.metadata.original_message.content,
-                sender: {
-                    userId: d.metadata.original_sender.id.toString(),
-                    username: d.metadata.original_sender.username,
-                    displayName: d.metadata.original_sender.username,
-                    profilePicture: '', // Not provided in event
-                    isVerified: false, // Worth checking?
-                    channelSlug: '' // Not provided in event, maybe should calculate from username?
-                }
-            } : undefined
-        };
-    }
-
-    private parseRewardRedeemedEvent(data: any): RewardRedeemedEvent {
-        const d = data as RewardRedeemedEventData;
-        return {
-            rewardTitle: d.reward_title,
-            userId: d.user_id,
-            channelId: d.channel_id,
-            username: d.username,
-            userInput: d.user_input,
-            rewardBackgroundColor: d.reward_background_color
-        };
-    }
-
-    private parseStreamHostedEvent(data: any): StreamHostedEvent {
-        const d = data as InboundStreamHostedEvent;
-        return {
-            user: {
-                userId: d.user.id.toString(),
-                username: d.user.username,
-                displayName: d.user.username,
-                profilePicture: '', // Not provided in event
-                isVerified: d.user.verified ? true : false,
-                channelSlug: '' // Not provided in event
-            },
-            numberOfViewers: d.message.numberOfViewers,
-            optionalMessage: d.message.optionalMessage,
-            createdAt: parseDate(d.message.createdAt)
-        };
     }
 
     private getBroadcaster(): KickUser {
